+++ packages/sdks/nextjs-sdk/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.12.8](https://github.com/descope/descope-js/compare/nextjs-sdk-0.12.7...nextjs-sdk-0.12.8) (2025-04-02)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.13.1`
+## [0.12.7](https://github.com/descope/descope-js/compare/nextjs-sdk-0.12.6...nextjs-sdk-0.12.7) (2025-04-02)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.30.0`
+* `react-sdk` updated to version `2.13.0`
+* `core-js-sdk` updated to version `2.41.0`
+* `web-component` updated to version `3.40.8`
+## [0.12.6](https://github.com/descope/descope-js/compare/nextjs-sdk-0.12.5...nextjs-sdk-0.12.6) (2025-04-02)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.12.6`
+## [0.12.5](https://github.com/descope/descope-js/compare/nextjs-sdk-0.12.4...nextjs-sdk-0.12.5) (2025-04-01)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.12.5`
 ## [0.12.4](https://github.com/descope/descope-js/compare/nextjs-sdk-0.12.3...nextjs-sdk-0.12.4) (2025-03-30)
 
 ### Dependency Updates